added unit test for app
