+++ babel.config.js
@@ -1,0 +1,7 @@
+
+module.exports = {
+  presets: [
+    '@babel/preset-env',
+    '@babel/preset-react',
+  ],
+};
