+++ client/src/components/App.jsx
@@ -1,0 +1,24 @@
+import React from 'react';
+
+// import Reviews from './components/Reviews.jsx';
+
+class App extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+
+    };
+  }
+
+  componentDidMount() {
+    // console.log(window.location.pathname);
+  }
+
+  render() {
+    return (
+      <div>hello</div>
+    );
+  }
+}
+
+export default App;
