+++ client/src/index.jsx
@@ -1,22 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import Reviews from './components/Reviews.jsx';
-
-class App extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-
-    };
-  }
-
-  render() {
-    return (
-      <div>hello</div>
-    );
-  }
-}
+import App from './components/App';
 
 
 ReactDOM.render(<App />, document.getElementById('app'));
