+++ package.json
@@ -38,7 +38,7 @@
     "eslint-plugin-react-hooks": "^2.5.1",
     "jest-environment-enzyme": "^7.1.1",
     "jest-enzyme": "^7.1.1",
-    "jest": "^25.5.2",
+    "jest": "^25.5.4",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11"
   }
